requirements.txt
cairosvg
lxml
numpy
pillow
svgpathtools